// packages/ERTP/src/types.d.ts
export type SetValue = any[];
// ...
export type Amount<K extends AssetKind = AssetKind> = {
  brand: Brand<K>;
  value: AssetValueForKind<K>;
};
// ...
export type AssetKind = 'nat' | 'set' | 'copySet' | 'copyBag';
export type AssetValueForKind<K extends AssetKind> = K extends 'nat'
  ? bigint
  : K extends 'set'
    ? SetValue
    : K extends 'copySet'
      ? CopySet
      : K extends 'copyBag'
        ? import('@endo/patterns').CopyBag<any>
        : never;
export type AssetKindForValue<V extends AmountValue> = V extends bigint
  ? 'nat'
  : V extends SetValue
    ? 'set'
    : V extends CopySet
      ? 'copySet'
      : V extends import('@endo/patterns').CopyBag<any>
        ? 'copyBag'
        : never;
/** @deprecated */
export type DisplayInfo<K extends AssetKind = AssetKind> = {
  /**
   * Tells the display software how many decimal places to move the decimal over
   * to the left, or in other words, which position corresponds to whole
   * numbers. We require fungible digital assets to be represented in integers,
   * in the smallest unit (i.e. USD might be represented in mill, a thousandth
   * of a dollar. In that case, `decimalPlaces` would be 3.) This property is
   * optional, and for non-fungible digital assets, should not be specified. The
   * decimalPlaces property should be used for _display purposes only_. Any
   * other use is an anti-pattern.
   */
  decimalPlaces?: number | undefined;
  /**
   * - the kind of asset, either AssetKind.NAT (fungible) or AssetKind.SET or
   *   AssetKind.COPY_SET (non-fungible)
   */
  assetKind: K;
};
// ...
export type Brand<K extends AssetKind = AssetKind> = {
  /** Should be used with `issuer.getBrand` to ensure an issuer and brand match. */
  isMyIssuer: (allegedIssuer: ERef<Issuer<AssetKind>>) => Promise<boolean>;
  getAllegedName: () => string;
  /** @deprecated look up in boardAux */
  getDisplayInfo: () => DisplayInfo<K>;
  getAmountShape: () => Pattern;
};
/**
 * Return true if the payment continues to exist.
 *
 * If the payment is a promise, the operation will proceed upon fulfillment.
 */
export type IssuerIsLive = (
  payment: ERef<Payment<AssetKind>>,
) => Promise<boolean>;
/**
 * Get the amount of digital assets in the payment. Because the payment is not
 * trusted, we cannot call a method on it directly, and must use the issuer
 * instead.
 *
 * If the payment is a promise, the operation will proceed upon fulfillment.
 */
export type IssuerGetAmountOf<K extends AssetKind> = (
  payment: ERef<Payment<AssetKind>>,
) => Promise<Amount<K>>;
// ...
export type IssuerBurn = (
  payment: ERef<Payment<AssetKind>>,
  optAmountShape?: Pattern,
) => Promise<Amount>;
/**
 * The issuer cannot mint a new amount, but it can create empty purses and
 * payments. The issuer can also transform payments (splitting payments,
 * combining payments, burning payments, and claiming payments exclusively). The
 * issuer should be gotten from a trusted source and then relied upon as the
 * decider of whether an untrusted payment is valid.
 */
export type Issuer<K extends AssetKind = AssetKind> = {
  /**
   * Get the Brand for this Issuer. The Brand indicates the type of digital
   * asset and is shared by the mint, the issuer, and any purses and payments of
   * this particular kind. The brand is not closely held, so this function
   * should not be trusted to identify an issuer alone. Fake digital assets and
   * amount can use another issuer's brand.
   */
  getBrand: () => Brand<K>;
  /** Get the allegedName for this mint/issuer */
  getAllegedName: () => string;
  /** Get the kind of MathHelpers used by this Issuer. */
  getAssetKind: () => AssetKind;
  /** @deprecated look up in boardAux */
  getDisplayInfo: () => DisplayInfo<K>;
  /** Make an empty purse of this brand. */
  makeEmptyPurse: () => Purse<K>;
  isLive: IssuerIsLive;
  getAmountOf: IssuerGetAmountOf<K>;
  burn: IssuerBurn;
};
// ...
export type DepositFacet = {
  /**
   * Deposit all the contents of payment into the purse that made this facet,
   * returning the amount. If the optional argument `optAmount` does not equal
   * the amount of digital assets in the payment, throw an error.
   *
   * If payment is a promise, throw an error.
   */
  receive: DepositFacetReceive;
};
export type PurseDeposit<K extends AssetKind> = (
  payment: Payment<K>,
  optAmountShape?: Pattern,
) => Amount<K>;
// ...
export type Purse<K extends AssetKind = AssetKind> = {
  /** Get the alleged Brand for this Purse */
  getAllegedBrand: () => Brand<K>;
  /** Get the amount contained in this purse. */
  getCurrentAmount: () => Amount<K>;
  /** Get a lossy notifier for changes to this purse's balance. */
  getCurrentAmountNotifier: () => LatestTopic<Amount<K>>;
  /**
   * Deposit all the contents of payment into this purse, returning the amount.
   * If the optional argument `optAmount` does not equal the amount of digital
   * assets in the payment, throw an error.
   *
   * If payment is a promise, throw an error.
   */
  deposit: PurseDeposit<K>;
  /** Return an object whose `receive` method deposits to the current Purse. */
  getDepositFacet: () => DepositFacet;
  /** Withdraw amount from this purse into a new Payment. */
  withdraw: (amount: Amount<K>) => Payment<K>;
  /**
   * The set of payments withdrawn from this purse that are still live. These
   * are the payments that can still be recovered in emergencies by, for
   * example, depositing into this purse. Such a deposit action is like
   * canceling an outstanding check because you're tired of waiting for it. Once
   * your cancellation is acknowledged, you can spend the assets at stake on
   * other things. Afterwards, if the recipient of the original check finally
   * gets around to depositing it, their deposit fails.
   *
   * Returns an empty set if this issuer does not support recovery sets.
   */
  getRecoverySet: () => CopySet<Payment<K>>;
  /**
   * For use in emergencies, such as coming back from a traumatic crash and
   * upgrade. This deposits all the payments in this purse's recovery set into
   * the purse itself, returning the total amount of assets recovered.
   *
   * Returns an empty amount if this issuer does not support recovery sets.
   */
  recoverAll: () => Amount<K>;
};
// ...
export type Payment<K extends AssetKind = AssetKind> = {
  /**
   * Get the allegedBrand, indicating the type of digital asset this payment
   * purports to be, and which issuer to use. Because payments are not trusted,
   * any method calls on payments should be treated with suspicion and verified
   * elsewhere.
   */
  getAllegedBrand: () => Brand<K>;
};